Extract models from server file
