+++ db/models/user.js
@@ -1,0 +1,19 @@
+const mongoose = require("mongoose")
+
+const UserSchema = mongoose.Schema({
+    name: {
+      type: String,
+      required: true
+    },
+    email: {
+      type: String,
+      required: true,
+      unique: true
+    },
+    password: {
+      type: String,
+      required: true
+    }
+  })
+
+  module.exports = mongoose.model("User", UserSchema)
